move everything into src folder
