+++ src/MessageAnalyzers/MessageAnalyzerInterface.ts
