+++ src/handlers/ResponseHandler.ts
@@ -5,7 +5,7 @@ import * as Greetings from '../MessageResponses/AppMentions/Greetings';
 import { GetRandomInt } from '../utils/Helpers';
 import Logger from '../utils/Logger';
 
-const config = require('../config.json');
+const config = require('../config/config.json');
 const bot_token = config.slack.bot_oauth_token || process.env.MANDO_BOT_TOKEN;
 
 /**List of happy emojis */
