+++ tsconfig.json
@@ -62,6 +62,7 @@
 
     /* Advanced Options */
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */,
-    "resolveJsonModule": true
-  }
+    "resolveJsonModule": true,
+  },
+  "exclude": ["node_modules", "./node_modules", "./node_modules/*"]
 }
